Move new classes to instancemute component
