+++ app/src/main/java/com/keylesspalace/tusky/components/instancemute/interfaces/InstanceActionListener.kt
@@ -1,4 +1,4 @@
-package com.keylesspalace.tusky.interfaces
+package com.keylesspalace.tusky.components.instancemute.interfaces
 
 interface InstanceActionListener {
     fun mute(mute: Boolean, instance: String, position: Int)
