+++ app/src/main/java/com/keylesspalace/tusky/di/ActivitiesModule.kt
@@ -16,6 +16,7 @@
 package com.keylesspalace.tusky.di
 
 import com.keylesspalace.tusky.*
+import com.keylesspalace.tusky.components.instancemute.InstanceListActivity
 import com.keylesspalace.tusky.components.report.ReportActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
